requirements.txt
argon2-cffi==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
ddt==1.2.1
Django==2.2.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
drf-yasg==1.15.0
entrypoints==0.3
Faker==0.9.1
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
openapi-codec==1.3.2
psycopg2==2.8.2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.15.96
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.25.2
